Add Factorial test class

+++ src/Factorial.java
@@ -1,5 +1,5 @@
 public class Factorial {
-    public static int factorial(int forNumber){
+    public static long factorial(int forNumber){
         if (forNumber == 0) return 1;
         else return(forNumber*factorial(forNumber-1));
     }
